update dotenv loading similar to job composer
load dotenv in config/boot
include in Configuration module
remove unnecessary guard clauses
load /etc.../.env and /etc.../.env.production files
