+++ config/examples/awesim/.env
